requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
gunicorn==21.2.0
